requirements.txt
# fastapi 相关插件（插件间接依赖未申明）
fastapi>=0.69.0
uvicorn>=0.14.0 # Python 3.6 只能安装0.16.0 否则 报错 AttributeError: module 'asyncio' has no attribute 'run'
gunicorn>=21.2.0

# fastflyer 相关插件（插件间接依赖未申明）
httptools>=0.1.2
fastkit>=0.1.12


# opentelemetry 相关插件
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-logging
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-httpx
opentelemetry-instrumentation-redis
opentelemetry-instrumentation-kafka-python
opentelemetry-instrumentation-pymysql
opentelemetry-instrumentation-sqlalchemy
opentelemetry-instrumentation-threading
